Improve snake-xenzia game mechanisms.
Tuned the random function to make food pixels appear more in the center of the screen in case some of the pixel rows go off screen. Also, now, if the snake is going in a particular direction, then it cannot immediately move in the opposite direction, ie, it cant move on top of itself.