Sorta fix mobile-snake bug.
Sort of fixed the mobile bug where the screen goes up and down. I've now forced the safari menu bar to show all the time but I'm starting to think its not feasible to change the weidth/height of the div in safari, this is more of a browser issue.